requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# PDF Processing
PyMuPDF==1.25.2
pdf2image==1.17.0
Pillow==11.1.0

# DOCX Processing
python-docx==1.1.2
mammoth==1.8.0

# LlamaParse (Primary Extraction)
llama-cloud-services==0.1.4

# Google Gemini (Secondary pass — diagram description + OCR verification)
google-genai==0.3.0

# Async + Queue (for scalability)
celery==5.4.0
redis==5.2.1
aiofiles==24.1.0
aiohttp==3.11.11

# Storage
google-cloud-storage==2.19.0

# Utilities
pydantic==2.10.5
pydantic-settings==2.7.1
loguru==0.7.3
python-dotenv==1.0.1
